+++ src/main/java/library/constant/Constant.java
@@ -1,9 +1,9 @@
-package library.constants;
+package library.constant;
 
-public final class Constants {
+public final class Constant {
 
     public static final long STANDARD_LOAN_PERIOD_WEEKS = 4;
     public static final long AGE_OF_MINORITY = 16;
 
-    private Constants() {}
+    private Constant() {}
 }
