+++ src/main/java/library/item/Book.java
@@ -1,7 +1,7 @@
-package library.items;
+package library.item;
 
-import library.constants.Constants;
-import library.customers.Customer;
+import library.constant.Constant;
+import library.customer.Customer;
 
 import java.time.LocalDate;
 
@@ -27,7 +27,7 @@ public class Book extends Item implements Borrowable {
     @Override
     public void loan(Customer customer, long loanWeeks) {
         if (!this.isForKids ||
-                LocalDate.now().minusYears(Constants.AGE_OF_MINORITY).isAfter(customer.getDateOfBirth())) {
+                LocalDate.now().minusYears(Constant.AGE_OF_MINORITY).isAfter(customer.getDateOfBirth())) {
             this.onLoan = true;
             this.loanExpiryDate = LocalDate.now().plusWeeks(loanWeeks);
         } else {
