+++ src/main/java/library/item/Borrowable.java
@@ -1,7 +1,7 @@
-package library.items;
+package library.item;
 
-import library.constants.Constants;
-import library.customers.Customer;
+import library.constant.Constant;
+import library.customer.Customer;
 
 import java.time.LocalDate;
 import java.util.UUID;
@@ -15,7 +15,7 @@ public interface Borrowable {
     void loan(Customer customer, long loanWeeks);
 
     default void loan(Customer customer) {
-        loan(customer, Constants.STANDARD_LOAN_PERIOD_WEEKS);
+        loan(customer, Constant.STANDARD_LOAN_PERIOD_WEEKS);
     }
 
     void returnFromLoan();
