+++ src/test/java/library/customer/CustomerTest.java
@@ -1,7 +1,8 @@
-package library.customers;
+package library.customer;
 
-import library.items.Book;
-import library.items.Borrowable;
+import library.builder.LibraryCustomerBuilder;
+import library.item.Book;
+import library.item.Borrowable;
 import org.junit.Before;
 import org.junit.Test;
